+++ src/components/Find/CombatantList.tsx
@@ -7,7 +7,7 @@ import {getDataBy} from 'data'
 import JOBS, {Role, ROLES} from 'data/JOBS'
 import {languageToEdition, PatchNumber, patchSupported} from 'data/PATCHES'
 import * as Errors from 'errors'
-import {Actor, ActorType} from 'fflogs2'
+import {Actor, ActorType} from 'fflogs'
 import {computed} from 'mobx'
 import {observer} from 'mobx-react'
 import AVAILABLE_MODULES from 'parser/AVAILABLE_MODULES'
