+++ src/parser/Conductor.ts
@@ -1,6 +1,6 @@
 import {getFflogsEvents} from 'api'
 import * as Errors from 'errors'
-import {Actor, Fight} from 'fflogs2'
+import {Actor, Fight} from 'fflogs'
 import {Report} from 'store/report'
 import {isDefined} from 'utilities'
 import AVAILABLE_MODULES from './AVAILABLE_MODULES'
