+++ src/parser/core/Parser.tsx
@@ -5,7 +5,7 @@ import ErrorMessage from 'components/ui/ErrorMessage'
 import {getReportPatch, languageToEdition} from 'data/PATCHES'
 import {DependencyCascadeError, ModulesNotFoundError} from 'errors'
 import type {Event} from 'events'
-import type {Actor, Fight, Pet} from 'fflogs2'
+import type {Actor, Fight, Pet} from 'fflogs'
 import React from 'react'
 import {Report} from 'store/report'
 import toposort from 'toposort'
