+++ src/parser/core/modules/BuffWindow.tsx
@@ -5,7 +5,7 @@ import {ActionLink} from 'components/ui/DbLink'
 import {RotationTable, RotationTableNotesMap, RotationTableTargetData} from 'components/ui/RotationTable'
 import {Action} from 'data/ACTIONS'
 import {Status} from 'data/STATUSES'
-import {BuffEvent, CastEvent} from 'fflogs2'
+import {BuffEvent, CastEvent} from 'fflogs'
 import _ from 'lodash'
 import Module, {dependency} from 'parser/core/Module'
 import GlobalCooldown from 'parser/core/modules/GlobalCooldown'
