+++ src/parser/core/modules/CooldownDowntime.tsx
@@ -2,7 +2,7 @@ import {t} from '@lingui/macro'
 import {Trans} from '@lingui/react'
 import {ActionLink} from 'components/ui/DbLink'
 import {Action} from 'data/ACTIONS'
-import {CastEvent} from 'fflogs2'
+import {CastEvent} from 'fflogs'
 import Module, {dependency} from 'parser/core/Module'
 import {Requirement, Rule} from 'parser/core/modules/Checklist'
 import React from 'react'
