+++ src/parser/core/modules/Invulnerability.tsx
@@ -1,6 +1,6 @@
 import Module, {dependency} from '../Module'
 import {Event} from 'events'
-import {Actor, ActorType} from 'fflogs2'
+import {Actor, ActorType} from 'fflogs'
 import React from 'react'
 import {Timeline, SimpleItem, SimpleRow} from './Timeline'
 import {AdditionalEvents, AdditionalEventQueries} from './AdditionalEvents'
