+++ src/parser/jobs/ast/modules/ArcanaTracking/ArcanaSuggestions.tsx
@@ -6,7 +6,7 @@ import {getDataBy} from 'data'
 import ACTIONS from 'data/ACTIONS'
 import JOBS from 'data/JOBS'
 import STATUSES from 'data/STATUSES'
-import {ActorType} from 'fflogs2'
+import {ActorType} from 'fflogs'
 import Module, {dependency} from 'parser/core/Module'
 import Combatants from 'parser/core/modules/Combatants'
 import {Timeline} from 'parser/core/modules/Timeline'
