+++ src/parser/jobs/ast/modules/ArcanaTracking/ArcanaTracking.tsx
@@ -2,7 +2,7 @@ import {t} from '@lingui/macro'
 import {getDataBy} from 'data'
 import ACTIONS from 'data/ACTIONS'
 import STATUSES from 'data/STATUSES'
-import {BuffEvent, CastEvent, DeathEvent, isCastEvent} from 'fflogs2'
+import {BuffEvent, CastEvent, DeathEvent, isCastEvent} from 'fflogs'
 import _ from 'lodash'
 import Module, {dependency} from 'parser/core/Module'
 import PrecastStatus from 'parser/core/modules/PrecastStatus'
