+++ src/parser/jobs/blm/modules/RotationWatchdog.tsx
@@ -7,7 +7,7 @@ import {ActionLink} from 'components/ui/DbLink'
 import {RotationTable} from 'components/ui/RotationTable'
 import {getDataBy} from 'data'
 import ACTIONS from 'data/ACTIONS'
-import {CastEvent, BaseEventFields} from 'fflogs2'
+import {CastEvent, BaseEventFields} from 'fflogs'
 import Module, {dependency} from 'parser/core/Module'
 import Checklist from 'parser/core/modules/Checklist'
 import Combatants from 'parser/core/modules/Combatants'
