+++ src/parser/jobs/dnc/modules/Esprit.tsx
@@ -9,7 +9,7 @@ import TimeLineChart from 'components/ui/TimeLineChart'
 import ACTIONS from 'data/ACTIONS'
 import JOBS from 'data/JOBS'
 import STATUSES from 'data/STATUSES'
-import {BuffEvent} from 'fflogs2'
+import {BuffEvent} from 'fflogs'
 import Module, {dependency} from 'parser/core/Module'
 import Combatants from 'parser/core/modules/Combatants'
 import {NormalisedDamageEvent} from 'parser/core/modules/NormalisedEvents'
