+++ src/parser/jobs/dnc/modules/Technicalities.tsx
@@ -9,7 +9,7 @@ import {RotationTable} from 'components/ui/RotationTable'
 import {getDataBy} from 'data'
 import ACTIONS from 'data/ACTIONS'
 import STATUSES from 'data/STATUSES'
-import {BuffEvent, CastEvent} from 'fflogs2'
+import {BuffEvent, CastEvent} from 'fflogs'
 import Module, {dependency} from 'parser/core/Module'
 import Combatants from 'parser/core/modules/Combatants'
 import {NormalisedApplyBuffEvent} from 'parser/core/modules/NormalisedEvents'
