+++ src/parser/jobs/war/modules/Gauge.tsx
@@ -3,7 +3,7 @@ import {Trans} from '@lingui/react'
 import ACTIONS from 'data/ACTIONS'
 import JOBS from 'data/JOBS'
 import STATUSES from 'data/STATUSES'
-import {CastEvent} from 'fflogs2'
+import {CastEvent} from 'fflogs'
 import {dependency} from 'parser/core/Module'
 import Combatants from 'parser/core/modules/Combatants'
 import {ComboEvent} from 'parser/core/modules/Combos'
